requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
fake-useragent==0.1.11
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
numpy==1.19.4
pandas==1.1.5
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.10.1
pyOpenSSL==20.0.0
python-dateutil==2.8.1
pytz==2020.4
queuelib==1.5.0
redis==3.5.3
redisbloom==0.4.0
requests==2.25.1
rmtest==0.7.0
Scrapy==2.4.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.21
Twisted @ file:///D:/chrome_download/Twisted-20.3.0-cp39-cp39-win_amd64.whl
urllib3==1.26.2
w3lib==1.22.0
zope.interface==5.2.0
